dev: code is separated based on clean architecture principles.
