+++ androidApp/src/main/java/com/abhishek/dongle/newsarticlesapp/android/di/ViewModelsModule.kt
@@ -1,6 +1,6 @@
 package com.abhishek.dongle.newsarticlesapp.android.di
 
-import com.abhishek.dongle.newsarticlesapp.article.ArticlesViewModel
+import com.abhishek.dongle.newsarticlesapp.presentation.ArticlesViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
